requirements.txt

reflex==0.8.17
psycopg2-binary